+++ src/types/plugins/rosterview/rosterview.d.ts
@@ -7,10 +7,15 @@ export default class RosterView extends CustomElement {
     initialize(): Promise<void>;
     model: Model;
     render(): import("lit-html").TemplateResult<1>;
-    showAddContactModal(ev: any): void;
-    syncContacts(ev: any): Promise<void>;
+    /** @param {MouseEvent} ev */
+    showAddContactModal(ev: MouseEvent): void;
+    /** @param {MouseEvent} [ev] */
+    syncContacts(ev?: MouseEvent): Promise<void>;
     syncing_contacts: boolean;
-    toggleRoster(ev: any): void;
+    /** @param {MouseEvent} [ev] */
+    toggleRoster(ev?: MouseEvent): void;
+    /** @param {MouseEvent} [ev] */
+    toggleFilter(ev?: MouseEvent): void;
 }
 import { CustomElement } from "shared/components/element.js";
 import { Model } from "@converse/skeletor";